fix: modal 스터디 이름 연결, 이모지 더 보기 오류 수정
